Moved example_bot to example dir
Moved the example bot file into an examples directory
